requirements.txt
python-dotenv>=1.0
pyyaml>=6.0
jsonschema>=4.17
httpx>=0.24
nextcord>=3.0
aiohttp>=3.9